requirements.txt
# Woody + Dashboard
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0
openai>=1.12.0
pydantic>=2.5.0

# OpenTelemetry
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-instrumentation-sqlite3>=0.48b0
opentelemetry-instrumentation-openai>=0.52.0; python_version >= "3.10"

# Natural language date parsing (handles "next monday", "tomorrow", etc.)
parsedatetime>=2.6

# Testing
pytest>=7.4.0

# Vector store (optional)
chromadb>=0.4.0

# Integrations
vobject>=0.9.7
twilio>=8.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0
